// src/diagram/control-panel/panels/types/constants.ts
export enum FoldButtons {
    Fold = 'fold',
}

export enum MoveButtons {
    Up = 'up',
    Down = 'down',
    Left = 'left',
    Right = 'right',
    UpRight = 'upRight',
    UpLeft = 'upLeft',
    DownLeft = 'downLeft',
    DownRight = 'downRight',
}

export enum ServiceButtons {
    Hide = 'hide',
    Fullscreen = 'fullscreen',
    Touch = 'touch',
}

export enum ZoomButtons {
    In = 'in',
    Out = 'out',
    Reset = 'reset',
}